refactor: share billing domain logic
