+++ apps/api/src/handlers/rates/resolve.test.ts
@@ -15,6 +15,7 @@ import { propertyConfig, propertyTestTimeout } from "@stll/property-testing";
 import type { Transaction } from "@/api/db/root";
 import type { SafeDb } from "@/api/db/safe-db";
 import { rateEntries, rateTables } from "@/api/db/schema";
+import { resolveRate } from "@/api/lib/billing-rates";
 import { createSafeDb, createScopedDb } from "@/api/db/scoped";
 import { toSafeId } from "@/api/lib/branded-types";
 import type { SafeId } from "@/api/lib/branded-types";
@@ -31,7 +32,7 @@ import type { TestIds } from "@/api/tests/security/rls-helpers";
 import { getTestDb, releaseTestDb } from "@/api/tests/security/test-utils";
 import type { TestDatabase } from "@/api/tests/security/test-utils";
 
-import resolveRateHandler, { resolveRate } from "./resolve";
+import resolveRateHandler from "./resolve";
 
 setDefaultTimeout(120_000);
