+++ apps/api/src/lib/billing-time.ts
@@ -1,0 +1,6 @@
+import { LIMITS } from "@/api/lib/limits";
+
+export const roundToBillingIncrement = (minutes: number): number => {
+  const increment = LIMITS.billingIncrementMinutes;
+  return Math.ceil(minutes / increment) * increment;
+};
